Stock-ms Render nuevo

+++ stock-ms/src/main/java/org/example/stockms/exception/ProductNotFoundException.java
@@ -6,10 +6,10 @@ import java.time.LocalDateTime;
 
 @Getter
 public class ProductNotFoundException extends RuntimeException {
-    private final Integer productId;
+    private final Long productId;
     private final LocalDateTime timestamp;
 
-    public ProductNotFoundException(Integer productId) {
+    public ProductNotFoundException(Long productId) {
         super("Producto con id " + productId + " no existe");
         this.productId = productId;
         this.timestamp = LocalDateTime.now();
